Align WS API with REST API
References to named events and actions in the REST API use the "name"
field and have one JSON object per named item. Previously, the WS API
inconsistently used the name parameter as the key for a JSON object
contained the remaining key-value pairs of the named item. This uses the
REST APIs approach throughout the WS API, losing the ability to
represent multiple named items in one API response.